added bootstrap and jquery and implemented a few basic components

+++ src/app/battle-selector/battle-selector.component.ts
@@ -18,7 +18,9 @@ export class BattleSelectorComponent implements OnInit {
 
   delete(idx: number) {
     console.log('pressed delete: ' + idx);
-    this.battleLibraryService.deleteBattleAndSelectNextOne(idx);
+    if (alert('Do you really want to delete ' + this.battleLibraryService.battles[idx].title + ' permanently?')) {
+      this.battleLibraryService.deleteBattleAndSelectNextOne(idx);
+    }
   }
 
   addBattle() {
